Modify travis instructions again

+++ run_docker.sh
@@ -1,4 +1,6 @@
-docker pull quay.io/pypa/manylinux1_x86_64
-docker run --rm -v `pwd`:/io quay.io/pypa/manylinux1_x86_64 /io/build-wheels.sh
-docker pull quay.io/pypa/manylinux1_i686
-docker run --rm -v `pwd`:/io quay.io/pypa/manylinux1_i686 linux32 /io/build-wheels.sh
+if [ -z $MULTIDICT_NO_EXTENSIONS ]; then
+    docker pull quay.io/pypa/manylinux1_x86_64
+    docker run --rm -v `pwd`:/io quay.io/pypa/manylinux1_x86_64 /io/build-wheels.sh
+    docker pull quay.io/pypa/manylinux1_i686
+    docker run --rm -v `pwd`:/io quay.io/pypa/manylinux1_i686 linux32 /io/build-wheels.sh
+fi
